\name{AssetCorr-package}
\alias{AssetCorr-package}
\alias{AssetCorr}
\docType{package}
\title{
AssetCorr
}
\description{ 
\packageDescription{AssetCorr}
}
\details{

The DESCRIPTION file:

\packageDESCRIPTION{AssetCorr}

\packageIndices{AssetCorr}
}


\author{
Maximilian Nagl [aut,cre],Yevhen Havrylenko [aut], Marius Pfeuffer [aut],  Matthias Fischer [aut],  Daniel Roesch [aut]

Maintainer: Maximilian Nagl <maximilian.nagl@ur.de>
}



\references{

De Servigny, A. and O. Renault: Default correlation: empirical evidence. Working Paper, Standard and Poor's: 90-94, 2003. Available at: \url{https://www.semanticscholar.org/paper/Default-correlation\%3A-empirical-evidence-Servigny-Renault/aae251436d0e3b489951c0d38463d71106755675}. Accessed: 04.05.2020

Duellmann, K. and M. Gehde-Trapp: Systematic risk in recovery rates: an empirical analysis of US corporate credit exposures. Bundesbank Series 2, Discussion Paper (2): 2004. Available at SSRN: \url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2793954}. Accessed: 01.05.2018

Duellmann, K., J. Kuell and M. Kunisch: Estimating asset correlations from stock prices or default rates- Which method is superior? Journal of Economic Dynamics and Control 34(11): 2341-2357, 2010

Efron, B. and R. J. Tibshirani: An introduction to the bootstrap. CRC press, 1994

Pfeuffer, M.,  M. Nagl, M. Fischer  and D. Roesch: Parameter Estimation and Bias Correction in the Vasicek Credit Portfolio Model.  Journal of Risk 22(4), 2020

Frei, C. and M. Wunsch: Moment Estimators for Autocorrelated Time Series and their Application to Default Correlations. Journal of Credit Risk 14: 1-29, 2018

Gordy, M. B.: A comparative anatomy of credit risk models. Journal of Banking &
Finance 24(1): 119-149, 2000

Gordy, M. B. and E. Heitfield: Small-sample estimation of models of portfolio credit risk. In Recent Advances in Financial Engineering: Proceedings of the KIER-TMU International Workshop on Financial Engineering, 2009: Otemachi, Sankei Plaza, Tokyo, 3-4 August
2009: 43-63, World Scientific, 2010

Hoese, S. and S. Huschens: Confidence intervals for asset correlations in the asymptotic single risk factor model. In Operations Research Proceedings 2010:  111-116, 2010

Kalkbrener, M. and A. Onwunta: Validating structural credit portfolio models. Model risk-identification, measurement and management. Risk Books, London: 233-261, 2010

Loeffler, G. The effects of estimation error on measures of portfolio credit risk. Journal of Banking & Finance 27(8): 1427-1453, 2003

Lucas, D. J.: Default correlation and credit analysis. The Journal of Fixed Income 4(4): 76-87, 1995

Meyer, C.: Estimation of intra-sector asset correlations. The Journal of Risk Model Validation 3(3): 47-79, 2009

Vasicek, O. A: The distribution of loan portfolio value. Risk 15(12): 160-162, 2002



}
